% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{Function to list files from a folder in owncloud}
\usage{
listFiles(username, password, uri = "cloud.josejesus.info/remote.php/webdav/",
  dir = "", file.type = "")
}
\arguments{
\item{username}{your user name
password your password
uri your owncloud address without http
dir  the directory or path contaning the sequences of directories where your files are
file.type the extension of the files. By default the function will list all files}
}
\description{
copyryig (c) 2016, José de Jesus Filho
last modified Nov, 2016
 
 This program is free software; you can redistribute it and/or
}
\details{
This functions allows you to list your files at your owncloud account
}
\examples{
listFiles(username="me",password="mypassword",uri="cloud.josejesus.info/remote.php/webdav/",dir="myfolder",file.type=".pdf")
}
\keyword{files}
\keyword{list}
\keyword{owncloud}

